Test for updating covariance on 2D linear system
Using the method I proposed in the AP Research Ideas Overleaf doc, this code updates a posteriori state covariance for a simple 2D linear system. It seems like the covariance it predicts is a multiple of the real covariance, and I'm still trying to figure out where that error comes from. But the fact they're almost proportional means it may be promising. It also seems like the specific vs used has a high impact on that proportionality constant